Update sp_Get_StaffCapability parameter and result types
Changed FromDate and ToDate parameters in sp_Get_StaffCapability from 'date' to 'datetime'. Updated ttgpljtnv and C_C1 properties in sp_Get_StaffCapability_Result from double to decimal for improved precision and consistency.

+++ NhatNam_web_new/NNworking/NNworking/Models/sp_Get_StaffCapability_Result.cs
@@ -15,13 +15,13 @@ namespace NNworking.Models
     {
         public string manv { get; set; }
         public Nullable<decimal> ttg { get; set; }
-        public Nullable<double> ttgpljtnv { get; set; }
+        public Nullable<decimal> ttgpljtnv { get; set; }
         public Nullable<double> ttgpljttc { get; set; }
         public Nullable<double> ttgpbjpbn { get; set; }
         public Nullable<double> ttgpbjttc { get; set; }
         public Nullable<decimal> ttgk { get; set; }
         public Nullable<double> C_B1 { get; set; }
-        public Nullable<double> C_C1 { get; set; }
+        public Nullable<decimal> C_C1 { get; set; }
         public Nullable<double> C_D1 { get; set; }
         public Nullable<double> C_E1 { get; set; }
         public Nullable<double> C_G1 { get; set; }
